Fix plugin.jar build issues
Added a minified plugin.jar that contains only one class, JSObject.  This new jar weighs in at 2KB (versus the 1MB version that ships with Java).   Placing this on the classpath first fixes the JavaFX namespace conflict, but allows compilation to succeed on all platforms without having to modify the project configuration.